feat: api(Auth 부분) 기초 세팅
