+++ app/types/auth.type.ts
@@ -1,5 +1,19 @@
-export interface UserInput {
-  email: string;
+export interface User {
+  id?: number;
+  email?: string;
   nickname?: string;
-  password: string;
+  profileImageUrl?: string;
+  createdAt?: string;
+  updatedAt?: string;
+}
+
+export interface UserData {
+  user?: User;
+  accessToken?: string;
+  refreshToken?: string;
+}
+
+export interface RequestToken {
+  accessToken?: string;
+  refreshToken?: string;
 }
